backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-decouple==2.1
djangorestframework==3.16.1
mysqlclient==2.2.7
python-dotenv==1.2.1
sqlparse==0.5.5
